+++ tests/assets/terser/main.js
@@ -1,0 +1,3 @@
+import { one, two } from "./numbers.js";
+
+console.log(one() + two());